requirements.txt
streamlit
pandas
pillow
gdown
patool